use a new index insted
